+++ deploy/debian.sh
@@ -1,0 +1,25 @@
+#!/bin/bash
+apt-get -y update
+apt-get -y install mc
+apt-get -y install wget
+apt-get -y install build-essential openssl libssl-dev pkg-config
+apt-get -y install python
+cd /usr/src
+wget http://nodejs.org/dist/v0.10.28/node-v0.10.28.tar.gz
+tar zxf node-v0.10.28.tar.gz
+cd node-v0.10.28
+./configure
+make
+make install
+echo 'deb http://downloads-distro.mongodb.org/repo/debian-sysvinit dist 10gen' | tee /etc/apt/sources.list.d/mongodb.list
+apt-get -y update
+apt-get -y --force-yes install mongodb-org
+service mongod start
+update-rc.d mongod defaults
+mkdir /impress
+cd /impress
+npm install mongodb
+npm install nodemailer
+npm install websocket
+npm install geoip-lite
+npm install impress